requirements.txt
discord.py
async-generator
async-timeout
nest-asyncio
captcha